Separated eventloop and websocket files
